chore: Tidy up dnspolicy controller code and files
Follow on from https://github.com/Kuadrant/kuadrant-operator/pull/937 to
cleanup controller files to bring them more inline with the current sotw
approach.

There are no logical changes in this commit, existing code is just being
moved.
