+++ controllers/dnspolicy_dnsrecords.go
@@ -2,6 +2,8 @@ package controllers
 
 import (
 	"fmt"
+	"net"
+	"strings"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	externaldns "sigs.k8s.io/external-dns/endpoint"
@@ -13,10 +15,13 @@ import (
 	"github.com/kuadrant/kuadrant-operator/api/v1alpha1"
 )
 
-var (
-	ErrNoRoutes    = fmt.Errorf("no routes attached to any gateway listeners")
-	ErrNoAddresses = fmt.Errorf("no valid status addresses to use on gateway")
+const (
+	LabelListenerReference = "kuadrant.io/listener-name"
 )
+
+func dnsRecordName(gatewayName, listenerName string) string {
+	return fmt.Sprintf("%s-%s", gatewayName, listenerName)
+}
 
 func desiredDNSRecord(gateway *gatewayapiv1.Gateway, clusterID string, dnsPolicy *v1alpha1.DNSPolicy, targetListener gatewayapiv1.Listener) (*kuadrantdnsv1alpha1.DNSRecord, error) {
 	rootHost := string(*targetListener.Hostname)
@@ -59,6 +64,60 @@ func desiredDNSRecord(gateway *gatewayapiv1.Gateway, clusterID string, dnsPolicy
 	}
 	dnsRecord.Spec.Endpoints = endpoints
 	return dnsRecord, nil
+}
+
+// GatewayWrapper is a wrapper for gateway to implement interface from the builder
+type GatewayWrapper struct {
+	*gatewayapiv1.Gateway
+	excludedAddresses []string
+}
+
+func NewGatewayWrapper(gateway *gatewayapiv1.Gateway) *GatewayWrapper {
+	return &GatewayWrapper{Gateway: gateway}
+}
+
+func (g *GatewayWrapper) GetAddresses() []builder.TargetAddress {
+	addresses := make([]builder.TargetAddress, len(g.Status.Addresses))
+	for i, address := range g.Status.Addresses {
+		addresses[i] = builder.TargetAddress{
+			Type:  builder.AddressType(*address.Type),
+			Value: address.Value,
+		}
+	}
+	return addresses
+}
+
+func (g *GatewayWrapper) RemoveExcludedStatusAddresses(p *v1alpha1.DNSPolicy) error {
+	g.excludedAddresses = p.Spec.ExcludeAddresses
+	newAddresses := []gatewayapiv1.GatewayStatusAddress{}
+	for _, address := range g.Gateway.Status.Addresses {
+		found := false
+		for _, exclude := range p.Spec.ExcludeAddresses {
+			//Only a CIDR will have  / in the address so attempt to parse fail if not valid
+			if strings.Contains(exclude, "/") {
+				_, network, err := net.ParseCIDR(exclude)
+				if err != nil {
+					return fmt.Errorf("could not parse the CIDR from the excludeAddresses field %w", err)
+				}
+				ip := net.ParseIP(address.Value)
+				// only check addresses that are actually IPs
+				if ip != nil && network.Contains(ip) {
+					found = true
+					break
+				}
+			}
+			if exclude == address.Value {
+				found = true
+				break
+			}
+		}
+		if !found {
+			newAddresses = append(newAddresses, address)
+		}
+	}
+	// setting this in memory only won't be saved to actual gateway
+	g.Status.Addresses = newAddresses
+	return nil
 }
 
 func buildEndpoints(clusterID, hostname string, gateway *gatewayapiv1.Gateway, policy *v1alpha1.DNSPolicy) ([]*externaldns.Endpoint, error) {
